requirements.txt
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.9
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.3
